+++ src/OString.php
@@ -49,6 +49,36 @@ class OString implements \Iterator, \ArrayAccess, \Serializable, \Countable, Con
         return $this->__toString();
     }
 
+
+    /**
+     * Alias to PHP function `explode`
+     *
+     * @param string $delim
+     * @param int    $limit
+     * @return OArray
+     * @throws \InvalidArgumentException
+     */
+    public function toOArray($delim = " ", $limit = null)
+    {
+        if (empty($this->string)) {
+            return new OArray();
+        }
+
+        if (empty($delim)) {
+            $delim = " ";
+        }
+
+        if (is_string($delim) || $delim instanceof OString) {
+            if (is_null($limit)) {
+                return new OArray(explode($delim, $this->string));
+            } elseif (is_integer($limit)) {
+                return new OArray(explode($delim, $this->string, $limit));
+            }
+            throw new \InvalidArgumentException("limit must be an integer");
+        }
+        throw new \InvalidArgumentException("delimiter must be a string");
+    }
+
     /**
      * @inheritdoc
      *