+++ CHANGELOG.md
@@ -1,9 +1,19 @@
 # Change Log
 
-## [0.1.1] - 2015-07-27
+## [0.2.0](https://github.com/ericpoe/ophp/tree/v0.2.0) - 2015-07-29
+OArray <--> String conversion
+
+* **Manual**
+    * Move Pipelining section to be closer to the top since it's a good example of what can be done
+    * Add OArray & OString- specific sections
+
+### Changes
+* OArray now has toOString() method. This is an alias to `implode`
+* OString now has toOArray() method. This is an alias to `explode`
+## [0.1.1](https://github.com/ericpoe/ophp/tree/v0.1.1) - 2015-07-27
 Documentation updates
 
-###Changes
+### Changes
 * Spacing requirements are now the same for all files
     * 4-space tabs
     * No empty space at the end of a line
@@ -16,7 +26,7 @@ Documentation updates
 * **README**
     * Add copy/paste one liner to require the library
 
-## [0.1.0] - 2015-07-26
+## [0.1.0](https://github.com/ericpoe/ophp/tree/v0.1.0) - 2015-07-26
 Initial release
 
 ### Contains